requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.7
django-rq==2.4.1
djangorestframework==3.12.4
pkg-resources==0.0.0
pytz==2021.1
redis==3.5.3
rq==1.8.0
sqlparse==0.4.1
